requirements.txt
fLask
flask-cors
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
pymysql